# app/interface/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the project of stampless team",
        "title": "Edu backend",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/users/login": {
            "post": {
                "description": "login an user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "login an user",
                "parameters": [
                    {
                        "description": "internal user email and password",
                        "name": "payloadBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.User"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "description": "Get information of logged in users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get information of logged in users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.User"
                        }
                    }
                }
            }
        },
        "/users/refresh": {
            "post": {
                "description": "refresh  for access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "refresh for access token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.User"
                        }
                    }
                }
            }
        },
        "/users/sign-up": {
            "post": {
                "description": "add an user to system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "add an internal user to system",
                "parameters": [
                    {
                        "description": "internal user email and role",
                        "name": "payloadBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UserAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "payload.UserAddRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "phonenumber",
                "rolecode",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                },
                "rolecode": {
                    "type": "string",
                    "enum": [
                        "user",
                        "manager",
                        "admin",
                        "superadmin"
                    ]
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "payload.UserLoginRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "presenter.User": {
            "type": "object",
            "properties": {
                "avatar_link": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}